search/requirements.txt
flask
flask_cors
google-cloud-firestore
